# src/main/kotlin/com/nevis/search/api/Dtos.kt
package com.nevis.search.api

import com.fasterxml.jackson.annotation.JsonPropertyOrder
import com.nevis.search.domain.Client
import com.nevis.search.domain.Document
import io.swagger.v3.oas.annotations.media.Schema
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.NotBlank
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Size
import java.time.OffsetDateTime
import java.util.UUID

// ---------- Requests ----------

@Schema(name = "CreateClientRequest")
data class CreateClientRequest(
    @field:NotBlank @field:Size(max = 255)
    @param:Schema(example = "John")
    val firstName: String?,

    @field:NotBlank @field:Size(max = 255)
    @param:Schema(example = "Doe")
    val lastName: String?,

    @field:NotBlank @field:Email @field:Size(max = 255)
    @param:Schema(example = "john.doe@neviswealth.com")
    val email: String?,

    @field:Size(max = 10_000)
    @param:Schema(example = "Retired surgeon, conservative risk profile, interested in ESG funds.")
    val description: String? = null,

    @field:Size(max = 20)
    @param:Schema(example = "[\"https://linkedin.com/in/johndoe\"]")
    val socialLinks: List<@NotBlank @Size(max = 2048) String>? = null,
)

@Schema(name = "CreateDocumentRequest")
data class CreateDocumentRequest(
    @field:NotBlank @field:Size(max = 255)
    @param:Schema(example = "Utility bill – March 2026")
    val title: String?,

    @field:NotBlank @field:Size(max = 200_000)
    @param:Schema(example = "Electricity bill for 12 Harbour View, London for the period 1–31 March 2026. Amount due £142.18.")
    val content: String?,
)

// ---------- Responses ----------

@Schema(name = "Client")
data class ClientResponse(
    val id: UUID,
    val firstName: String,
    val lastName: String,
    val email: String,
    val description: String?,
    val socialLinks: List<String>,
    val createdAt: OffsetDateTime,
) {
    companion object {
        fun from(c: Client) = ClientResponse(
            id = c.id,
            firstName = c.firstName,
            lastName = c.lastName,
            email = c.email,
            description = c.description,
            socialLinks = c.socialLinks,
            createdAt = c.createdAt,
        )
    }
}

@Schema(name = "Document")
data class DocumentResponse(
    val id: UUID,
    val clientId: UUID,
    val title: String,
    val content: String,
    @param:Schema(description = "Short summary of the content. Generated by Claude when configured, otherwise the leading sentences.")
    val summary: String?,
    @param:Schema(description = "Related search terms attached at write time (thesaurus and/or Claude). Searchable.")
    val keywords: List<String>,
    val createdAt: OffsetDateTime,
) {
    companion object {
        fun from(d: Document) = DocumentResponse(
            id = d.id,
            clientId = d.client.id,
            title = d.title,
            content = d.content,
            summary = d.summary,
            keywords = d.keywords,
            createdAt = d.createdAt,
        )
    }
}

// ---------- Search ----------

enum class SearchType { CLIENT, DOCUMENT }

@Schema(
    name = "SearchResult",
    description = "One hit. `type` says which of `client` / `document` is populated.",
    oneOf = [ClientSearchResult::class, DocumentSearchResult::class],
    discriminatorProperty = "type",
)
sealed interface SearchResult {
    val type: String

    @get:Schema(description = "Relevance in [0, 1]; results are ordered by it descending.", example = "0.87")
    val score: Double

    @get:Schema(description = "Which fields produced the match, e.g. [\"email\"] or [\"semantic\"].")
    val matchedOn: List<String>
}

@Schema(name = "ClientSearchResult")
@JsonPropertyOrder("type", "score", "matched_on", "client")
data class ClientSearchResult(
    override val score: Double,
    override val matchedOn: List<String>,
    val client: ClientResponse,
) : SearchResult {
    @get:Schema(allowableValues = ["client"])
    override val type: String get() = "client"
}

@Schema(name = "DocumentSearchResult")
@JsonPropertyOrder("type", "score", "matched_on", "document")
data class DocumentSearchResult(
    override val score: Double,
    override val matchedOn: List<String>,
    val document: DocumentResponse,
) : SearchResult {
    @get:Schema(allowableValues = ["document"])
    override val type: String get() = "document"
}